Fixed issue with unit tests and included auth in nested routers
