+++ app/api/auth/jwt_bearer.py
@@ -36,3 +36,6 @@ class JWTBearer(HTTPBearer):
         if decode_token:
             return True
         return False
+
+
+authenticate = JWTBearer()
