+++ app/api/routers/user.py
@@ -11,7 +11,7 @@ from services import UserService, get_db
 from auth.jwt_bearer import JWTBearer
 
 
-router = APIRouter(prefix="/user", tags=["user"], dependencies=[Depends(JWTBearer())])
+router = APIRouter(prefix="/user", tags=["user"], dependencies=[Depends(JWTBearer)])
 
 
 @router.get("/{user_id}", response_model=user_schemas.User, status_code=200)
