+++ app/api/tests/conftest.py
@@ -1,6 +1,12 @@
 import pytest
 import sqlalchemy
+from unittest.mock import patch
+from auth.jwt_bearer import JWTBearer
+from fastapi.testclient import TestClient
+
 import models
+from main import app
+from services import get_db
 from .db.session import engine
 
 
@@ -20,3 +26,22 @@ def setup():
     with engine.connect() as conn:
         conn.execute(sqlalchemy.text("DROP SCHEMA IF EXISTS core"))
         conn.commit()
+
+
+@pytest.fixture()
+def client(session):
+    def override_get_db():
+        yield session
+
+    def mock_authenticate():
+        with patch.object(JWTBearer, "__call__") as mock_call:
+            # Implement the desired behavior of the mocked __call__ method
+            mock_call.return_value = "mocked_token"
+
+            yield
+
+    app.dependency_overrides[get_db] = override_get_db
+    app.dependency_overrides[JWTBearer] = mock_authenticate
+    yield TestClient(app)
+    del app.dependency_overrides[get_db]
+    del app.dependency_overrides[JWTBearer]
