+++ app/api/tests/test_model.py
@@ -3,7 +3,8 @@ from unittest.mock import patch, Mock
 import schemas.model as model_schemas
 import schemas.model_details as model_details_schemas
 import schemas.user as user_schemas
-from .db.session import client, session
+from .db.session import session
+from .conftest import client
 from utils.model_builder import ModelBuilder
