+++ app/api/tests/test_pool.py
@@ -1,6 +1,7 @@
 import schemas.pool as pool_schemas
 import schemas.user as user_schemas
-from .db.session import client, session
+from .db.session import session
+from .conftest import client
 
 USER_ROUTE = "/user"
 POOL_ROUTE = "/pool"
